Added everything useful from 0-6000
There was a lot of tat and stuff that wouldn't work for what we needed.
When I revisit this I will go through and add more that are just the top
half of the player and so on but for release this is fine as is. Still
have to look through 6000-7000 as I've seen some small things in there.
Also /handsup needs to be set to PD commands.