/help fully working and optimised
I removed the /FAQ and have added that to the help menu itself. There
are some questions left empty as we dont have a /map yet but suggest to
use /n and the forums for help. But is all working. I'll continue to
update this if people udate the thread.