Finished off LSNN
Decided not to place cams as an inventory item since they're exclusive
to LSNN members and not necessary for /give, etc. Also added a clear
focus/add focus streamer for cams and a working chopper cam that updates
rotation every second. Having issues with 'headline', it's not
displaying any text.